narrative-audio-system/task3_transcription/requirements.txt
openai-whisper==20231117
torch==2.2.1
numpy==1.26.4